+++ packages/plugins/MousePosition/src/components/index.ts
@@ -1,0 +1,1 @@
+export { default as MousePosition } from './MousePosition.vue'